% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_data}
\alias{get_data}
\title{get_data()}
\usage{
get_data(option)
}
\description{
scraping information of today's covid19
and seperate continent, country and total data
}
\examples{
get_data("country")

}
