% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_data_csv}
\alias{get_data_csv}
\title{get_data_csv()}
\usage{
get_data_csv(option)
}
\arguments{
\item{option}{character. is country or continent}
}
\description{
scraping information of today's covid19
and return csv file
}
\examples{
get_data_csv("continent")
}
