% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_data_sas}
\alias{get_data_sas}
\title{get_data_sas()}
\usage{
get_data_sas(option)
}
\arguments{
\item{option}{character. is country or continent}
}
\description{
scraping information of today's covid19
and return .sas file
}
\examples{
get_data_sas("continent")
}
