% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_data_spss}
\alias{get_data_spss}
\title{get_data_spss()}
\usage{
get_data_spss(option)
}
\arguments{
\item{option}{character. is country or continent}
}
\description{
scraping information of today's covid19
and return .sav file
}
\examples{
get_data_spss("country")
}
