% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_data_stata}
\alias{get_data_stata}
\title{get_data_stata()}
\usage{
get_data_stata(option)
}
\arguments{
\item{option}{character. is country or continent}
}
\description{
scraping information of today's covid19
and return .stata file
}
\examples{
get_data_stata("country")
}
