% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_data_excel}
\alias{get_data_excel}
\title{get_data_excel()}
\usage{
get_data_excel(option)
}
\arguments{
\item{option}{character. is country or continent}
}
\description{
scraping information of today's covid19
and return xlsx file
}
\examples{
get_data_excel("continent")

}
